requirements.txt
altgraph==0.17.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
gnupg==2.3.1
idna==3.7
packaging==24.1
psutil==6.0.0
pycparser==2.22
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyYAML==6.0.1
requests==2.32.3
setuptools==72.1.0
urllib3==2.2.2
